Fixing dtype issue in siglip layer

+++ keras_hub/src/models/siglip/siglip_layers.py
@@ -73,7 +73,7 @@ class SigLIPVisionEmbedding(layers.Layer):
             # Let the backend determine the int dtype. For example, tf
             # requires int64 for correct device placement, whereas jax and torch
             # don't.
-            dtype=int,
+            dtype="int32",
             trainable=False,
             name="position_ids",
         )
@@ -197,7 +197,7 @@ class SigLIPTextEmbedding(layers.Layer):
             # Let the backend determine the int dtype. For example, tf
             # requires int64 for correct device placement, whereas jax and torch
             # don't.
-            dtype=int,
+            dtype="int32",
             trainable=False,
             name="position_ids",
         )
